Add `maxdepth` parameter to `get_labels` function (#3460)
Add a `maxdepth` parameter to the `get_labels` built-in function that
controls the maximum depth of `get_labels`' recursive dependency search.
The default is -1, meaning "no limit". The main use case is to help
limit the collection of `cc:inc:` labels to direct dependencies in the
cc-rules plugin, since headers used by transitive (library) dependencies
shouldn't be required at compilation time.

This is an alternative to (and is in fact a superset of the
functionality offered by) the `transitive` parameter:
`transitive=True` is equivalent to `maxdepth=-1`, and
`transitive=False` is equivalent to `maxdepth=0`. The `transitive`
parameter remains for backwards compatibility, although it could now be
removed in a future major release.