Optionally traverse `srcs` and `deps` for run-time dependencies
There are circumstances under which a target might want to inherit the
run-time dependencies of certain build-time dependencies, because the
target's output is built in such a way that its dependencies don't have
their own run-time dependencies resolved. An example of this is the
shell-rules plugin's `sh_binary` rule: a `sh_binary` is a zip file of
its dependencies (`deps`) concatenated onto a shell script preamble that
extracts those dependencies into a temporary directory during
initialisation. If one of those `deps` happens to have its own run-time
dependencies, Please will not resolve them, and they will therefore not
be built before the target runs (and will not be sent to the remote
worker along with the target's output).

Add two parameters to `build_rule` - `runtime_deps_from_srcs` and
`runtime_deps_from_deps` - that indicate to Please that run-time
dependencies should additionally be collected from the target's `srcs`
and `deps` respectively. This allows build definitions to inherit the
run-time dependencies of their sources and/or build-time dependencies if
they know they will be needed when running the target they generate.